Added second model with curved adiabats

+++ main.py
@@ -75,29 +75,55 @@ def main():
         sys.exit()
     if sim == 'FSSH':
         if scan:
-            from input import gen_sh_input, write_input
-            for W in W_vals:
-                for B in B_vals:
-                    for A in A_vals:
-                        for p_vec in p_vec_list:
-                            input_data, filename = gen_sh_input(A, B, W, N, p_vec[0], p_vec[1], rescale_method)
-                            write_input(filename, input_data)
-                            with open(filename) as f:
-                                for line in f:
-                                    line1 = line.replace(" ", "")
-                                    line1 = line1.rstrip('\n')
-                                    name, value = line1.split("=")
-                                    exec(str(line), globals())
-                            copyfile(filename, 'inputfile.tmp')
-                            from fssh import runSim, genviz
-                            if not (os.path.exists(calcdir)):
-                                os.mkdir(calcdir)
-                            runSim()
-                            genviz()
-                            os.remove('inputfile.tmp')
-                            del runSim
-                            del genviz
-                            del sys.modules['fssh']
+            if model==1:
+                from input import gen_sh_input, write_input
+                for W in W_vals:
+                    for B in B_vals:
+                        for A in A_vals:
+                            for p_vec in p_vec_list:
+                                input_data, filename = gen_sh_input(A, B, W, N, p_vec[0], p_vec[1], rescale_method)
+                                write_input(filename, input_data)
+                                with open(filename) as f:
+                                    for line in f:
+                                        line1 = line.replace(" ", "")
+                                        line1 = line1.rstrip('\n')
+                                        name, value = line1.split("=")
+                                        exec(str(line), globals())
+                                copyfile(filename, 'inputfile.tmp')
+                                from fssh import runSim, genviz
+                                if not (os.path.exists(calcdir)):
+                                    os.mkdir(calcdir)
+                                runSim()
+                                genviz()
+                                os.remove('inputfile.tmp')
+                                del runSim
+                                del genviz
+                                del sys.modules['fssh']
+            if model==2:
+                from input import gen_sh_input_2, write_input
+                for W in W_vals:
+                    for B in B_vals:
+                        for A in A_vals:
+                            for alpha in alpha_vals:
+                                for p_vec in p_vec_list:
+                                    input_data, filename = gen_sh_input_2(alpha,A, B, W, N, p_vec[0], p_vec[1], rescale_method)
+                                    write_input(filename, input_data)
+                                    with open(filename) as f:
+                                        for line in f:
+                                            line1 = line.replace(" ", "")
+                                            line1 = line1.rstrip('\n')
+                                            name, value = line1.split("=")
+                                            exec(str(line), globals())
+                                    copyfile(filename, 'inputfile.tmp')
+                                    from fssh_2 import runSim, genviz
+                                    if not (os.path.exists(calcdir)):
+                                        os.mkdir(calcdir)
+                                    runSim()
+                                    genviz()
+                                    os.remove('inputfile.tmp')
+                                    del runSim
+                                    del genviz
+                                    del sys.modules['fssh_2']
         else:
             # run FSSH dynamics
             from fssh import runSim, genviz
